Digital Quiz System
This was my final year project at the university. It is developed in React Native, Expo, and Firebase. This app contains 2 portals i.e for Students & Teachers. Teachers are responsible for maintaining the class, courses, pool of questions, and Students. They are able to perform CRUD operations and assign the quizzes to the Students at a specific date & time. While the Students receive the notification about the given course and able to conduct the quiz. The result generating process is automized Teachers & Students get to see the scores of the particular Student